renamed SubmitReducer
